+++ app/lib/main.dart
@@ -59,8 +59,8 @@ class _OudsApplicationState extends State<OudsApplication> {
                     ? TextDirection.rtl // If the language is RTL, use TextDirection.rtl
                     : TextDirection.ltr, // Otherwise, use TextDirection.ltr
                 child: OudsThemeConfigModel(
-                  button: OudsButtonConfig(rounded: true),
-                  tag: OudsTagConfig(rounded: false),
+                  button: OudsButtonConfig(rounded: themeController.onBorderRadiusButtonState),
+                  tag: OudsTagConfig(rounded: themeController.onBorderRadiusTagState),
                   textInput: OudsTextInputConfig(rounded: false),
                   child: OudsTheme(
                     themeContract: themeController.currentTheme,