+++ ouds_core/lib/components/button/internal/ouds_button_border_modifier.dart
@@ -13,7 +13,6 @@
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_loading_modifier.dart';
 import 'package:ouds_core/components/button/ouds_button.dart';
-import 'package:ouds_theme_contract/config/component/ouds_button_config.dart';
 import 'package:ouds_theme_contract/config/ouds_theme_config_model.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
 
@@ -47,6 +46,8 @@ class OudsButtonBorderModifier {
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
         return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongEnabled, width: theme.componentsTokens(context).button.borderWidthDefault) : BorderSide.none;
+      case OudsButtonHierarchy.brand:
+        return BorderSide.none;
       case OudsButtonHierarchy.minimal:
         return BorderSide.none;
       case OudsButtonHierarchy.negative:
@@ -63,6 +64,8 @@ class OudsButtonBorderModifier {
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
         return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongHover, width: theme.componentsTokens(context).button.borderWidthDefaultInteraction) : BorderSide.none;
+      case OudsButtonHierarchy.brand:
+        return BorderSide.none;
       case OudsButtonHierarchy.minimal:
         return BorderSide.none;
       case OudsButtonHierarchy.negative:
@@ -79,6 +82,8 @@ class OudsButtonBorderModifier {
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
         return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongPressed, width: theme.componentsTokens(context).button.borderWidthDefaultInteraction) : BorderSide.none;
+      case OudsButtonHierarchy.brand:
+        return BorderSide.none;
       case OudsButtonHierarchy.minimal:
         return BorderSide.none;
       case OudsButtonHierarchy.negative:
@@ -96,6 +101,8 @@ class OudsButtonBorderModifier {
     switch (hierarchy) {
       case OudsButtonHierarchy.strong:
         return onColoredSurface ? BorderSide(color: theme.componentsTokens(context).buttonMono.colorBorderStrongDisabled, width: theme.componentsTokens(context).button.borderWidthDefault) : BorderSide.none;
+      case OudsButtonHierarchy.brand:
+        return BorderSide.none;
       case OudsButtonHierarchy.minimal:
         return BorderSide.none;
       case OudsButtonHierarchy.negative: