+++ ouds_core/lib/components/button/internal/ouds_button_style_modifier.dart
@@ -13,6 +13,7 @@
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_background_modifier.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_border_modifier.dart';
+import 'package:ouds_core/components/button/internal/ouds_button_control_state.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_foreground_modifier.dart';
 import 'package:ouds_core/components/button/internal/ouds_button_padding_modifier.dart';
 import 'package:ouds_core/components/button/ouds_button.dart';
@@ -23,24 +24,15 @@ class OudsButtonStyleModifier {
     BuildContext context, {
     required OudsButtonHierarchy hierarchy,
     required OudsButtonLayout layout,
+    bool? border,
     OudsButtonStyle? style,
-    required bool isPressed,
+    OudsButtonControlState? buttonState,
   }) {
-    double iconSize;
-    if (layout == OudsButtonLayout.iconOnly) {
-      iconSize = OudsTheme.of(context).componentsTokens(context).button.sizeIconOnly;
-    } else if (layout == OudsButtonLayout.iconAndText) {
-      iconSize = OudsTheme.of(context).componentsTokens(context).button.sizeIcon;
-    } else {
-      iconSize = 0.0;
-    }
     return ButtonStyle(
-      backgroundColor: OudsButtonBackgroundModifier.resolveBackgroundColor(context, hierarchy, style, isPressed),
-      foregroundColor: OudsButtonForegroundModifier.resolveForegroundColor(context, hierarchy, style, isPressed),
-      iconColor: OudsButtonForegroundModifier.resolveForegroundColor(context, hierarchy, style, isPressed),
+      backgroundColor: OudsButtonBackgroundModifier.resolveBackgroundColor(context, hierarchy, style, buttonState),
+      foregroundColor: OudsButtonForegroundModifier.resolveForegroundColor(context, hierarchy, style, buttonState),
       splashFactory: NoSplash.splashFactory,
       overlayColor: WidgetStateProperty.all(Colors.transparent),
-      iconSize: WidgetStateProperty.all<double>(iconSize),
       textStyle: WidgetStateProperty.all<TextStyle>(
         TextStyle(
             fontSize: OudsTheme.of(context).fontTokens.sizeLabelLarge,